Task-15-companies-support correct after Radek's comments

+++ src/main/java/pl/fc/invoicing/controllers/CompanyController.java
@@ -1,7 +1,6 @@
 package pl.fc.invoicing.controllers;
 
 import java.util.List;
-import java.util.NoSuchElementException;
 import java.util.Optional;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
@@ -13,6 +12,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import pl.fc.invoicing.dto.CompanyDto;
 import pl.fc.invoicing.dto.CompanyListDto;
+import pl.fc.invoicing.exceptions.handlers.IdNotFoundException;
 import pl.fc.invoicing.services.CompanyService;
 
 @Slf4j
@@ -49,14 +49,13 @@ public class CompanyController implements CompanyControllerApi {
 
     @Override
     public ResponseEntity<Void> delete(@PathVariable UUID id) {
-        try {
+        if (companyService.getById(id).isPresent()) {
             companyService.delete(id);
-        } catch (NoSuchElementException e) {
             log.debug("Deleting invoice by id: " + id);
             return ResponseEntity.status(204).build();
+        } else {
+            throw new IdNotFoundException("Company id: " + id + " not found.");
         }
-        log.debug("Cannot delete invoice by id: " + id);
-        return ResponseEntity.noContent().build();
     }
 
     @Override
